Use repo URI that comes back from github
This kind of solves some captitalization issues. For instance, Github
will respond to requests to github.com/datomic/codeq, even though the
real URI is github.com/Datomic/codeq.  The URI info given back will
always be with the prefered capitalization, however.  This change uses
the info that comes back, so that we won't reimport repositories because
of different capitalization.  The issue still exists for local repos if
they have a different capitalization in the remote uri.