Add Pygments highlight option
If h flag is passed to codeq, it will import a :code/highlight attribute
along with each text segment it imports.  It uses the Pygments
highlighter to do this, and outputs to HTML.  The pygmentize script must
be on the command line, or in the directory the jar is being excecuted.
Highlighting support is independent of the analyzer; it should work with
future analyzers as long as they import the :code/text attribute.